ARMODServers/requirements.txt
aliyun-python-sdk-core==2.13.33
aliyun-python-sdk-kms==2.14.0
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crcmod==1.7
cryptography==3.4.7
Django==3.2
django-redis==4.12.1
djangorestframework==3.12.4
docopt==0.6.2
idna==2.10
importlib-metadata==3.10.1
itsdangerous==1.1.0
jmespath==0.10.0
kombu==5.0.2
mysqlclient==2.0.3
oss2==2.14.0
pipreqs==0.4.10
prompt-toolkit==3.0.18
pycparser==2.20
pycryptodome==3.10.1
PyMySQL==1.0.2
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
supervisor==4.2.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
yarg==0.1.9
zipp==3.4.1
